% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_peta.R
\name{opa}
\alias{opa}
\title{Check points outside their polygons}
\usage{
opa(ldmark = NULL, poly_map = NULL)
}
\arguments{
\item{ldmark}{A geojson file containing project name, project description, iddesa, project type, etc.}

\item{poly_map}{A data frame (geojson file) containing region information,
typically including province, municipality/regency, district and village information.}
}
\value{
The function is called for check points outside their polygons and will save.
  its result in Excel report.
}
\description{
Check points outside its SLS polygon by iddesa.
}
\examples{
\dontrun{
# Make sure your working directory contains landmark and digital map files
# opa(ldmark = "my points.geojson", poly_map = "my final sls.geojson")
}
}
