% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_peta.R
\name{org_peta}
\alias{org_peta}
\title{Organize Scanned Map Files}
\usage{
org_peta(kodekab = NULL, datawil = NULL)
}
\arguments{
\item{kodekab}{A character string representing the district/municipality
code to filter files in this directory.}

\item{datawil}{A data frame (csv/geojson file) containing region information,
typically including province, district/municipality, and village information.}
}
\value{
The function is called for its side effect of organizing files.
  It also prints messages to the console about the progress and results.
}
\description{
Organizing scanned map files based on the SLS codes
  and creating directories that correspond to each SLS code.
}
\examples{
\dontrun{
# Make sure your working directory contains scanned map files
# starting with kodekab.
# setwd("path/to/your/maps")
# org_peta(kodekab = "3273", datawil = "my final sls.geojson")
}
}
