% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_peta.R
\name{rename_peta}
\alias{rename_peta}
\title{Rename Scanned Map Files using OCR}
\usage{
rename_peta(kodekab = NULL)
}
\arguments{
\item{kodekab}{A character string representing the municipality/regency
code to verify the renamed files.}
}
\value{
The function is called for its side effect of renaming files.
  It also prints messages to the console about the progress and results.
}
\description{
Automatically renames and rotates multiple JPEG map files in the current
  working directory based on a 14-digit SLS code found within the image.
}
\examples{
\dontrun{
# To run this function, make sure you have .jpg files in your
# working directory.
# setwd("path/to/your/maps")
# rename_peta(kodekab = "3273")
}
}
