% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_peta.R
\name{sbr_out_desa}
\alias{sbr_out_desa}
\title{Mengidentifikasi usaha/perusahaan hasil profiling SBR yang berada di luar desa}
\usage{
sbr_out_desa(
  dir.titik = NULL,
  dir.desa.sls = NULL,
  target = c("gc", "profiling")
)
}
\arguments{
\item{dir.titik}{Direktori file csv hasil scraping di Website/Aplikasi Matchapro.}

\item{dir.desa.sls}{Direktori file geojson peta desa/SLS terbaru (gunakan peta hasil Pemetaan Wilkerstat SE2026)}
}
\value{
Fungsi ini akan mengembalikan file excel dan geojson yang berisi.
  usaha-usaha/perusahaan-perusahaan yang berada di luar desa masing-masing.
}
\description{
Mengidentifikasi usaha/perusahaan hasil profiling SBR yang berada di luar desa berdasarkan iddesa.
}
\examples{
\dontrun{
# sbr_out_desa(dir.titik = "direktori_usaha_full_all_columns_2026.csv", dir.desa.sls = "Final_Desa_202415205.geojson")
}
}
