% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tpt0525.R
\docType{data}
\name{tpt0525}
\alias{tpt0525}
\title{Data Tingkat Pengangguran Terbuka (TPT) Indonesia}
\format{

}
\source{
BPS Indonesia, 2025
}
\usage{
tpt0525
}
\description{
Dataset ini berisi TPT Indonesia dalam satuan persen (%) yang dirilis oleh BPS RI setiap bulan Februari dan Agustus.
}
\examples{
data(tpt0525)
head(tpt0525)
}
\keyword{datasets}
